Refactors models page and adds new hero component
Updates the models page to feature the Leopard 7, streamlining the presentation and navigation.

Implements a new hero component with improved visuals and user engagement.

Adds image assets and updates components to use webp format for improved performance.

Adds a suppressHydrationWarning prop to the body element to resolve a hydration issue.